requirements.txt
# Core
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Document generation
python-docx>=1.1.0

# Browser automation
playwright>=1.44.0

# Web scraping
beautifulsoup4>=4.12.0

# Dashboard
streamlit>=1.35.0
pandas>=2.2.0
